-discarded tiles are now shown -adjusted code to draw any n by m board -fixed bug where the tiles that are lying down wouldn't adjust to the height set properly. Changed the parameters of Obj::draw so that the tile would be translated using floats, not ints

+++ Obj.cpp
@@ -199,7 +199,7 @@ void Obj::loadObj()
     glEndList();
 }
 
-void Obj::draw(int x, int y, int z, float rot_x, float rot_y, float rot_z)
+void Obj::draw(float x, float y, float z, float rot_x, float rot_y, float rot_z)
 {
     glPushMatrix();
     glTranslatef(x, y, z);
